requirements.txt
django>=5.0
djangorestframework
gunicorn
psycopg2-binary
dj-database-url
python-dotenv
whitenoise
reportlab
